Added way more validation and a Validator

+++ App/Backend/src/main/java/com/cryptotrading/cryptotrading/services/impl/HoldingServiceImpl.java
@@ -6,6 +6,7 @@ import com.cryptotrading.cryptotrading.domain.dto.response.HoldingDto;
 import com.cryptotrading.cryptotrading.domain.dto.response.ViewHoldingsResponseDto;
 import com.cryptotrading.cryptotrading.mappers.Mapper;
 import com.cryptotrading.cryptotrading.services.HoldingService;
+import com.cryptotrading.cryptotrading.util.Validator;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
@@ -20,15 +21,33 @@ public class HoldingServiceImpl implements HoldingService {
     private final HoldingDao holdingDao;
     private final Mapper<Holding, HoldingDto> holdingMapper;
 
-    public HoldingServiceImpl(HoldingDao holdingDao, Mapper<Holding, HoldingDto> holdingMapper) {
+    private final Validator validator;
+
+    public HoldingServiceImpl(HoldingDao holdingDao, Mapper<Holding, HoldingDto> holdingMapper, Validator validator) {
         this.holdingDao = holdingDao;
         this.holdingMapper = holdingMapper;
+        this.validator = validator;
     }
 
 
     @Override
     public void buyHolding(UUID userId, String symbol, BigDecimal amount) {
         Holding holding = holdingDao.findByUserIdAndSymbol(userId, symbol);
+
+        if(!validator.isUUIDValid(userId))
+        {
+            return;
+        }
+
+        if(validator.isStringForNullOrEmpty(symbol))
+        {
+            return;
+        }
+
+        if(!validator.isPositive(amount))
+        {
+            return;
+        }
 
         if(holding == null) {
             createHolding(userId, symbol, amount);
@@ -38,6 +57,11 @@ public class HoldingServiceImpl implements HoldingService {
 
         BigDecimal newAmount = holding.getAmount().add(amount);
 
+        if(!validator.isPositive(newAmount))
+        {
+            return;
+        }
+
         holding.setAmount(newAmount);
 
         holdingDao.update(holding);
@@ -45,20 +69,32 @@ public class HoldingServiceImpl implements HoldingService {
 
     @Override
     public void sellHolding(UUID userId, String symbol, BigDecimal amount) {
+        if(!validator.isUUIDValid(userId)) {
+            return;
+        }
+
         Holding holding = holdingDao.findByUserIdAndSymbol(userId, symbol);
 
         if(holding == null) {
             return;
         }
 
-        if(holding.getAmount().compareTo(amount) < 0) {
+        if(!validator.isPositive(amount)) {
+            return;
+        }
+
+        if(validator.isZero(amount)) {
             return;
         }
 
         BigDecimal newAmount = holding.getAmount().subtract(amount);
 
-        if(newAmount.compareTo(BigDecimal.ZERO) == 0) {
+        if(validator.isZero(newAmount)) {
             holdingDao.delete(holding.getId());
+            return;
+        }
+
+        if(!validator.isPositive(newAmount)) {
             return;
         }
 
@@ -70,13 +106,25 @@ public class HoldingServiceImpl implements HoldingService {
     }
 
     @Override
-    public Holding getByUserIdAndSmybol(UUID userId, String symbol) {
+    public Holding getByUserIdAndSymbol(UUID userId, String symbol) {
+        if(!validator.isUUIDValid(userId)) {
+            return null;
+        }
+
+        if(validator.isStringForNullOrEmpty(symbol)) {
+            return null;
+        }
+
         return holdingDao.findByUserIdAndSymbol(userId, symbol);
     }
 
     @Override
-    public void deleteUserHoldings(UUID userId)
-    {
+    public void deleteUserHoldings(UUID userId) {
+        if(!validator.isUUIDValid(userId))
+        {
+            return;
+        }
+
         holdingDao.deleteUserHoldings(userId);
     }
 
@@ -85,7 +133,21 @@ public class HoldingServiceImpl implements HoldingService {
     {
         ViewHoldingsResponseDto result =  new ViewHoldingsResponseDto();
 
+        if(!validator.isUUIDValid(userId)) {
+            result.setStatus(false);
+            result.setErrorMessage("Invalid User");
+
+            return result;
+        }
+
         List<Holding> holdingList = holdingDao.getUserHoldings(userId);
+
+        if(holdingList == null) {
+            result.setStatus(false);
+            result.setErrorMessage("There was an error getting holdings");
+
+            return result;
+        }
 
         List<HoldingDto> holdingDtoList = result.getHoldings();
 
@@ -100,6 +162,18 @@ public class HoldingServiceImpl implements HoldingService {
 
     @Override
     public void createHolding(UUID userId, String symbol, BigDecimal amount) {
+        if(!validator.isUUIDValid(userId)) {
+            return;
+        }
+
+        if(validator.isStringForNullOrEmpty(symbol)) {
+            return;
+        }
+
+        if(!validator.isPositive(amount)) {
+            return;
+        }
+
         Holding holding = Holding.builder()
                 .userId(userId)
                 .symbol(symbol)
